ClusteredLowRankSolver.jl

ClusteredLowRankSolver.jl provides a primal-dual interior point method for solving clustered low-rank semidefinite programs. This can primarily be used for (semidefinite) programs with polynomial inequality constraints, which can be rewritten in terms of sum-of-squares polynomials. See the manual for a more detailed description of the problems and the usage of the solver.

ClusteredLowRankSolver.Block — Type

Block(l::Any[, r::Int, s::Int])

Specifies a block corresponding to the positive semidefinite variable l.

Specifying r,s makes the Block correspond to the r,s subblock of the variable l.

ClusteredLowRankSolver.ClusteredLowRankSDP — Type

ClusteredLowRankSDP(sos::LowRankSOSProblem[, as_free])

Define a ClusteredLowRankSDP based on the LowRankSOSProblem sos.

The PSD variables defined by the keys as_free will be modelled as extra free variables, with extra constraints to ensure that they are equal to the entries of the PSD variables. Keyword arguments: prec (default: precision(BigFloat)) - the precision of the result verbose (default: false) - print progress to the standard output

ClusteredLowRankSolver.Constraint — Type

Constraint(constant, matrixcoeff, freecoeff, samples[, scalings])

Models a polynomial constaint of the form constant(x) = ∑block < block, matrixcoeffblock > + ∑freevar freecoefffreevar*freevar sampled on the elements of samples.

Arguments: constant::MPolyElem matrixcoeff::Dict(Block, LowRankMatPol) freecoeff::Dict(Any, MPolyElem) samples::Vector{Vector} scalings::Vector

ClusteredLowRankSolver.LowRankMatPol — Type

LowRankMat(eigenvalues::Vector, rightevs::Vector{Vector}[, leftevs::Vector{Vector}])

The matrix ∑_i eigenvalues[i]rightevs[i]leftevs[i]'.

If leftevs is not specified, use leftevs = rightevs. The elements of the Vectors are

ClusteredLowRankSolver.Objective — Type

Objective(constant, matrixcoeff::Dict{Block, Matrix}, freecoeff::Dict)

The objective for the LowRankSOSProblem

ClusteredLowRankSolver.approximatefekete — Method

approximatefekete(basis, samples) -> basis, samples

Compute approximate fekete points based on samples and a corresponding orthogonal basis.

This preserves a degree ordering of basis if present.

ClusteredLowRankSolver.calculate_res_d! — Method

Compute the dual residue d = c - <A_*, Y> - By

ClusteredLowRankSolver.check_pd_feasibility — Method

Check primal and dual feasibility

ClusteredLowRankSolver.compute_S_integrated! — Method

Compute S, integrated with the precomputing of the bilinear pairings

ClusteredLowRankSolver.compute_T_decomposition! — Method

Compute the decomposition of [S B; B^T 0]

ClusteredLowRankSolver.compute_dual_error — Method

Compute the dual error

ClusteredLowRankSolver.compute_dual_objective — Method

Compute the dual objective <C,Y> + <b,y> + constant

ClusteredLowRankSolver.compute_duality_gap — Method

Compute the duality gap (primalobj - dualobj)/max{1, |primalobj+dualobj|}

ClusteredLowRankSolver.compute_error — Method

Compute the error (max abs(d_ij)) of an ArbMatrix

ClusteredLowRankSolver.compute_error — Method

Compute the error (max abs (P_ij)) of a blockdiagonal matrix

ClusteredLowRankSolver.compute_primal_error — Method

Compute the primal error

ClusteredLowRankSolver.compute_primal_objective — Method

Compute the primal objective <c, x> + constant

ClusteredLowRankSolver.compute_residual_R! — Method

Compute the residual R, with or without second order term

ClusteredLowRankSolver.compute_residuals! — Method

Compute the residuals P,p and d.

ClusteredLowRankSolver.compute_search_direction! — Method

Compute the search directions, using a precomputed decomposition

ClusteredLowRankSolver.compute_step_length — Method

Compute the step length min(γ α(M,dM), 1), where α is the maximum number step to which keeps M+α(M,dM) dM positive semidefinite

ClusteredLowRankSolver.compute_weighted_A! — Method

Set initialmatrix to ∑i ai Ai

ClusteredLowRankSolver.distribute_weights_swapping — Method

Distribute the weights over n (about) equal-sized bins by swapping between bins with high total weight and low total weight.

ClusteredLowRankSolver.dot_c — Method

Compute <c,x> where c is distributed over constraints

ClusteredLowRankSolver.solvesdp — Function

solvesdp(sdp; kwargs...)

Solve the clustered SDP with low-rank constraint matrices.

Solve the following sdp: 'max/min ∑j <C^j, Y^j> + b^T y s.t. <A^j,Y^j> + B^j y = c^j Y^j ⪰ 0' where we optimize over the free variables 'y' and the PSD block matrices 'Y^j = diag(Y^{j,1}, ..., Y^{j,Lj})', and '<A^j, Y^j> = (<Ap^j, Y^j>){p=1,...,Pj}'. The matrices A^jp have the same block structure as Y^j. Every A^{j,l} can have several equalsized blocks A^{j,l}[r,s]. The smallest blocks have a low rank structure.

Important keyword arguments which are not clear from the name: - gamma: the step length reduction; a maximum step length of α reduces to a step length of max(gammaα,1) - beta(in)feasible: the amount mu is tried to be reduced by in each iteration, for (in)feasible solutions - omegap/d: the starting matrix variable for the primal/dual is omega_p/dI

ClusteredLowRankSolver.terminate — Method

Determine whether the main loop should terminate or not

ClusteredLowRankSolver.trace_A — Method

Compute the vector <A*,Z> = Tr(A* Z) for one or all constraints