ClusteredLowRankSolver.jl

ClusteredLowRankSolver.jl provides a primal-dual interior point method for solving clustered low-rank semidefinite programs. This can be used for (semidefinite) programs with polynomial inequality constraints, which can be rewritten in terms of sum-of-squares polynomials. See the manual for a detailed description of the problems and the usage of the solver.

Citation

If you use ClusteredLowRankSolver.jl in work that results in a publication, consider citing

D. de Laat and N.M. Leijenhorst, Solving Clustered Low-Rank Semidefinite Programs arising from Polynomial Optimization, arXiv:

Clustered Low-Rank Semidefinite Programs

\[\begin{aligned} \min \quad & \sum_j \langle Y^j, C^j \rangle + \langle y, b\rangle \\ \text{s.t.} \quad & \langle Y^j, A^j_* \rangle + B^T y = c \\ & Y^j \succeq 0 \end{aligned}\]

Documentation

ClusteredLowRankSolver.Block — Type

Block(l::Any[, r::Int, s::Int])

Specifies a block corresponding to the positive semidefinite variable l.

Specifying r,s makes the Block correspond to the r,s subblock of the variable l.

ClusteredLowRankSolver.ClusteredLowRankSDP — Type

ClusteredLowRankSDP(sos::LowRankSOSProblem[, as_free])

Define a ClusteredLowRankSDP based on the LowRankSOSProblem sos.

The PSD variables defined by the keys as_free will be modelled as extra free variables, with extra constraints to ensure that they are equal to the entries of the PSD variables. Keyword arguments: prec (default: precision(BigFloat)) - the precision of the result verbose (default: false) - print progress to the standard output

ClusteredLowRankSolver.Constraint — Type

Constraint(constant, matrixcoeff, freecoeff, samples[, scalings])

Models a polynomial constaint of the form constant = ∑block < block, matrixcoeff[block] > + ∑freevar freecoeff[freevar]*freevar with samples.

Arguments: constant::MPolyElem matrixcoeff::Dict(Block, LowRankMatPol) freecoeff::Dict(Any, MPolyElem) samples::Vector{Vector} scalings::Vector

ClusteredLowRankSolver.LowRankMatPol — Type

LowRankMat(eigenvalues::Vector, rightevs::Vector{Vector}[, leftevs::Vector{Vector}])

The matrix ∑_i eigenvalues[i]rightevs[i]leftevs[i]'.

If leftevs is not specified, use leftevs = rightevs. The elements of the Vectors are

ClusteredLowRankSolver.Objective — Type

Objective(constant, matrixcoeff::Dict{Block, Matrix}, freecoeff::Dict)

The objective for the LowRankSOSProblem

ClusteredLowRankSolver.approximatefekete — Method

approximatefekete(basis, samples) -> basis, samples

Compute approximate fekete points based on samples and a corresponding orthogonal basis. The basis consists of sampled polynomials, sampled at samples

This preserves a degree ordering of basis if present.

ClusteredLowRankSolver.calculate_res_d! — Method

Compute the dual residue d = c - <A_*, Y> - By

ClusteredLowRankSolver.check_pd_feasibility — Method

Check primal and dual feasibility

ClusteredLowRankSolver.compute_S_integrated! — Method

Compute S, integrated with the precomputing of the bilinear pairings

ClusteredLowRankSolver.compute_T_decomposition! — Method

Compute the decomposition of [S B; B^T 0]

ClusteredLowRankSolver.compute_dual_error — Method

Compute the dual error

ClusteredLowRankSolver.compute_dual_objective — Method

Compute the dual objective <C,Y> + <b,y> + constant

ClusteredLowRankSolver.compute_duality_gap — Method

Compute the duality gap (primalobj - dualobj)/max{1, |primalobj+dualobj|}

ClusteredLowRankSolver.compute_error — Method

Compute the error (max abs(d_ij)) of an ArbMatrix

ClusteredLowRankSolver.compute_error — Method

Compute the error (max abs (P_ij)) of a blockdiagonal matrix

ClusteredLowRankSolver.compute_primal_error — Method

Compute the primal error

ClusteredLowRankSolver.compute_primal_objective — Method

Compute the primal objective <c, x> + constant

ClusteredLowRankSolver.compute_residual_R! — Method

Compute the residual R, with or without second order term

ClusteredLowRankSolver.compute_residuals! — Method

Compute the residuals P,p and d.

ClusteredLowRankSolver.compute_search_direction! — Method

Compute the search directions, using a precomputed decomposition

ClusteredLowRankSolver.compute_step_length — Method

Compute the step length min(γ α(M,dM), 1), where α is the maximum number step to which keeps M+α(M,dM) dM positive semidefinite

ClusteredLowRankSolver.compute_weighted_A! — Method

Set initialmatrix to ∑i ai Ai

ClusteredLowRankSolver.distribute_weights_swapping — Method

Distribute the weights over n (about) equal-sized bins by swapping between bins with high total weight and low total weight.

ClusteredLowRankSolver.dot_c — Method

Compute <c,x> where c is distributed over constraints

ClusteredLowRankSolver.solvesdp — Function

solvesdp(sdp; kwargs...)

Solve the clustered SDP with low-rank constraint matrices.

Solve the following sdp:

max/min ∑_j <C^j, Y^j> + b^T y
s.t.    <A_*^j,Y^j> + B^j y = c^j
        Y^j ⪰ 0

where we optimize over the free variables y and the PSD block matrices Y^j = diag(Y^{j,1}, ..., Y^{j,L_j}), and <A_*^j, Y^j> = (<A_p^j, Y^j>)_{p=1,...,P_j}. The matrices A^j_p have the same block structure as Y^j. Every A^{j,l} can have several equal-sized blocks A^{j,l}[r,s]. The smallest blocks have a low rank structure.

Keyword arguments: - prec (default: precision(BigFloat)): the precision used - gamma (default: 0.9): the step length reduction; a maximum step length of α reduces to a step length of max(gamma*α,1) - beta_(in)feasible (default: 0.1 (0.3)): the amount mu is tried to be reduced by in each iteration, for (in)feasible solutions - omega_p/d (default: 10^10): the starting matrix variable for the primal/dual is omega_p/d*I - maxiterations (default: 500): the maximum number of iterations - duality_gap_threshold (default: 10^-15): how near to optimal the solution needs to be - primal/dual_error_threshold (default:10^-30): how feasible the primal/dual solution needs to be - max_complementary_gap (default: 10^100): the maximum of <X,Y>/#rows(X) allowed - need_primal_feasible/need_dual_feasible (default: false): terminate when the solution is primal/dual feasible - verbose (default: true): print information after every iteration if true - step_length_threshold (default: 10^-7): the minimum step length allowed - initial_solutions (default: []): if x,X,y,Y are given, use that instead of omega_p/d * I for the initial solutions

ClusteredLowRankSolver.terminate — Method

Determine whether the main loop should terminate or not

ClusteredLowRankSolver.trace_A — Method

Compute the vector <A*,Z> = Tr(A* Z) for one or all constraints